backend/requirements/base.txt
# Core Django
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
django-db-geventpool==4.0.1

# Caching & Task Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Authentication & JWT
djangorestframework-simplejwt==5.3.2
PyJWT==2.8.1

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# File Storage (AWS S3)
boto3==1.34.15
django-storages==1.14.2

# Data Validation & Serialization
marshmallow==3.20.1
python-dateutil==2.8.2
pytz==2023.3
python-decouple==3.8

# Monitoring & Logging
sentry-sdk==1.38.0
python-json-logger==2.0.7

# Utilities
python-slugify==8.0.1
Pillow==10.1.0
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2

# Security
cryptography==41.0.7
